LPD-70691 Refactor - Create shared PhoneNumberInput in object-js-components-web
This also moves LocalizablePhoneNumber into a single PhoneNumber.tsx since this component is a lot more simplified now so it'll be easier to manage both the non-localized and localized versions of the phone number input.
