+++ modules/apps/object/object-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/js/index.ts
@@ -13,5 +13,4 @@ export {default as Attachment} from './Attachment/Attachment';
 export {default as AutoIncrement} from './AutoIncrement/AutoIncrement';
 export {default as MultiselectPicklist} from './MultiselectPicklist/MultiselectPicklist';
 export {default as ObjectRelationship} from './ObjectRelationship/ObjectRelationship';
-export {default as LocalizablePhoneNumber} from './PhoneNumber/LocalizablePhoneNumber';
 export {default as PhoneNumber} from './PhoneNumber/PhoneNumber';
